添加质检完成字段，添加文件编号

+++ src/main/java/com/deepsoft/haolifa/dao/repository/FileRecordMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface FileRecordMapper {
-    long countByExample(FileRecordExample example);
+    int countByExample(FileRecordExample example);
 
     int deleteByExample(FileRecordExample example);
